Package and method refactor
