+++ src/com/sourdoughsoftware/Game.java
@@ -1,6 +1,7 @@
 package com.sourdoughsoftware;
 
 import org.xml.sax.SAXException;
+import com.sourdoughsoftware.utility.*;
 
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
@@ -35,7 +36,7 @@ public class Game {
 //        get room description
         String userName = parser.prompt("Enter your name adventurer\n>> ");
         p1 = new Player(userName, 0);
-        String roomDescription = getRoomDescription(p1.getPlayerRoomID(), rooms);
+        String roomDescription = Utilities.getRoomDescription(p1.getPlayerRoomID(), rooms);
         System.out.println(roomDescription);
 
         do {
@@ -48,7 +49,7 @@ public class Game {
             String[] userCommands = parser.promptAction(">> ");
             if (userCommands[0].equals("go")) {
                 System.out.println(RoomChange.changeRoom(p1.getPlayerRoomID(), userCommands[1], rooms, p1));
-                roomDescription = getRoomDescription(p1.getPlayerRoomID(), rooms);
+                roomDescription = Utilities.getRoomDescription(p1.getPlayerRoomID(), rooms);
             }
             if (userCommands[0].equalsIgnoreCase("look") ||
                     userCommands[0].equalsIgnoreCase("show") ||
@@ -76,7 +77,7 @@ public class Game {
                 continue;
             }
             if (Verbs.getAllVerbs().contains(userCommands[0].toLowerCase()) && Nouns.getAllNouns().contains(userCommands[1].toLowerCase())) {
-                String message = processInteraction(rooms, userCommands);
+                String message = parser.processInteraction(rooms, userCommands, p1);
                 // output verb interaction message or a failure message
                 System.out.println(message);
                 continue;
@@ -86,59 +87,6 @@ public class Game {
         } while (!gameOver);
     }
 
-    private String processInteraction(List<Room> rooms, String[] userCommands) {
-        String message = "You try to " + userCommands[0].toLowerCase() + " the " + userCommands[1].toLowerCase() + " but nothing happens.";
-        // get room items against player location
-        List<Item> itemsToCheck = getPlayerRoomItems(p1.getPlayerRoomID(), rooms);
-
-        if (userCommands[0].equalsIgnoreCase("take") ||
-                userCommands[0].equalsIgnoreCase("get")) {
-            // Get player item
-            Item playerItem = new Item();
-            for (Item item : itemsToCheck) {
-                if (item.getName().equalsIgnoreCase(userCommands[1])) {
-                    playerItem = item;
-                    break;
-                }
-            }
-            message = TakeItem.takeItem(p1, rooms, playerItem);
-        } else {
-            // check item for verb interaction map
-            for (Item item : itemsToCheck) {
-                if (item.getVerbInteraction().containsKey(userCommands[0].toLowerCase()) &&
-                        item.getName().equalsIgnoreCase(userCommands[1])) {
-                    message = item.getVerbInteraction().get(userCommands[0].toLowerCase());
-                    break;
-                }
-            }
-        }
-        return message;
-    }
-
-    private List<Item> getPlayerRoomItems(Integer playerRoomID, List<Room> rooms) {
-        List<Item> specificItems = new ArrayList<>();
-        for (Room room : rooms) {
-            if (playerRoomID.equals(room.getRoomID())) {
-                specificItems = room.getRoomItems();
-            }
-        }
-        return specificItems;
-    }
-
-    private String getRoomDescription(Integer playerLocation, List<Room> rooms) {
-        String result = "";
-        for (Room room : rooms) {
-            if (playerLocation.equals(room.getRoomID())) {
-                if (room.getDescription() == null) {
-                    room.setDescription("Room " + room.getRoomID() + " is missing a description!");
-                    result = room.getDescription();
-                } else {
-                    result = room.getDescription();
-                }
-            }
-        }
-        return result;
-    }
 
     public void setGameOver(boolean gameOver) {
         this.gameOver = gameOver;
