+++ src/com/sourdoughsoftware/utility/Look.java
@@ -1,4 +1,8 @@
-package com.sourdoughsoftware;
+package com.sourdoughsoftware.utility;
+
+import com.sourdoughsoftware.Item;
+import com.sourdoughsoftware.Player;
+import com.sourdoughsoftware.Room;
 
 import java.util.List;
