+++ src/com/sourdoughsoftware/utility/Nouns.java
@@ -1,10 +1,8 @@
-package com.sourdoughsoftware;
+package com.sourdoughsoftware.utility;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.List;
-import java.util.stream.Stream;
 
 public enum Nouns {
     // Story specific items
